WIP Generate structs for all event arguments.
This commit generates structs and Environment implementations for all types of event arguments, similar to the treatment of state arguments and state variables.

Previously, only event arguments were handled in this way, while all other event arguments were stored in a HashMap associated with the event.

This fixes that inconsistency while making it easier to provide uniform runtime access to all kinds of arguments/variables.

Currently, this build is broken since the new structs are generated, but not used correctly.