Tons of manual clippy fixes.
Most of these are pretty non-controversial and make the code better.

The only ones I have misgivings about are that I removed the common suffixes from the enum variants in `scanner.rs`. I think this makes the code better, but it touches a lot of lines of code...

There were similar warnings in `ast.rs`, but I just suppressed these since the type names are longer (so the qualified variant names are longer) and it would touch even more lines of code.

I suppressed several frequent warnings in the non-Rust visitors and added TODOs. I figure we can fix these as we go, when these backends get more attention in the future.