Codegen for new interface (CI will fail)
This is a mostly complete extension of the code generator to implement the new and significantly expanded runtime interface. All that's left is transitions, which are the trickiest bit, but I've got a plan.

Compilation of generated code for tests will fail in the current state, and the tests will still need to be updated once code generation is finished.