WIP codegen for new context/event representation
This is a mostly complete implementation of the codegen for the refactored state context and event representations, needed to support the extended runtime interface. Most of the codegen for the new runtime interface itself is also done, but there are likely still gaps there.

Unfortunately, I realized that my refactoring of how domain variables are represented is going to break almost all existing code and make the interface for simple cases worse... so I'm going to back those changes out, which is the reason for this WIP commit.

Essentially all end-to-end tests are disabled in this commit due to the problem described above and since there are likely other issues still lurking in the new codegen stuff.