Major refactor of runtime interface (CI will fail)
This is a major refactor of the runtime interface which makes it *much* easier to program against in a generic way. The most obvious changes are (1) the interface is no longer split into sync/unsync variants, (2) the runtime interface no longer requires a lifetime annotation on the machine type.

The type annotations required on callback arguments have gotten a little bit harrier... I really wish Rust was better at inferring the types of closure arguments--in this case all the info is obviously there to do it.

The runtime interface and its demos and tests have been updated and are working, but the code generator and integration tests have not been updated yet, so CI is expected to fail.