Weaken bounds on associated types.
This change was needed so that the Send and Sync traits will propagate to structs that contain values of these associated types, when applicable. Unfortunately, Rust can't automatically propagate the associated type bounds, which means they need to be propagated everywhere that the Machine trait is used. Quite annoying from an ergonomic perspective...

Here's the RFC to add the ability to automatically propagate trait bounds to Rust in the future: https://github.com/rust-lang/rust/issues/52662