Document Bug #18: Domain variable lines incorrectly mapped
The transpiler is mapping domain variable initialization lines
(e.g., 'var data = None' on line 37) to multiple Python lines
within the __init__ method. This causes the debugger to incorrectly
show Frame line 37 when stepping through various parts of the
constructor code.

This is a transpiler source map generation bug that needs to be
fixed in the Frame transpiler itself.